live-backend/requirements.txt
alembic==1.0.10
aniso8601==6.0.0
asn1crypto==1.2.0
attrs==19.1.0
bcrypt==3.1.6
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.4.0
cryptography==2.8
elasticsearch==7.1.0
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Excel==0.0.7
flask-marshmallow==0.10.1
Flask-Migrate==2.5.1
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
geographiclib==1.49
geopy==1.20.0
googlemaps==3.1.3
googletrans==2.4.0
HiYaPyCo==0.4.14
HTMLParser==0.0.2
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jsonschema==3.0.1
lml==0.0.9
Mako==1.0.10
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
nltk==3.4.5
numpy==1.16.4
openapi-spec-validator==0.2.8
openpyxl==3.0.3
pandas==0.25.0
psycopg2-binary==2.8.4
pycparser==2.19
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.8
PyJWT==1.7.1
pymongo==3.8.0
PyMySQL==0.9.3
pyodbc==4.0.27
pyOpenSSL==19.1.0
pypyodbc==1.3.2
pyrsistent==0.15.2
PySocks==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
PyYAML==5.2
requests==2.22.0
schedule==0.6.0
six==1.13.0
SQLAlchemy==1.3.3
swagger-ui==0.1.2
swagger-ui-bundle==0.0.6
texttable==1.6.2
urllib3==1.25.7
Werkzeug==0.15.4
win-inet-pton==1.1.0
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
gunicorn==20.0.4
